**Problem**: psychopy/tools/linebreak_class.py used in linebreak.py
Current implementation stores all linebreak codes in a massive dictionary (300K+ lines). This works for normal running (~0.6 s on my setup) but slows initialization (imports) dramatically when debugging (~60 s). The slow start up times when debugging are general issue replicated both on Windows and Macs.

**Solution:** as suggested by @peircej, import inside the method only if line breaking is "aux14". Solves problem for everyone who uses "default" line breaks.