@peircej 2022-07-25
We could, of course, use both methods I guess. Western language users don't need anything to happen and users of uax14 could get a csv file (we'd need to cache it so we don't reload the file on each call) so their debugger doesn't do so much work 

Longer term, it feels like we could do something to get rid (or at least cut down) this massive file - there's a lot of redundancy in there, like hundreds of thousands of entries for the same thing where we might simply specify a range instead. But we'd need to check whether that was slower on each get() call than the simple dict lookup

@hsogo any thoughts?