@alexander-pastukhov 2022-07-25
@peircej Yes, we can reduce it to under a couple of thousands of sequences. That was my first idea because it is indeed very redundant. But, perhaps, this could be a way to construct the dictionary (once), instead of using if-elif in the function itself. This way we can keep fast look up but cut the file size by two orders of magnitude. What do you think?