@hsogo 2022-07-25
I'm sorry I left this code for a long time.  I haven't tested how fast (slow) if-elif implementation is.  I'll try it.