@hsogo 2022-07-26
I wrote a function that converts codes to line-breaking classes by a long chain of if-sentences.  The code was 7002 lines long.  I measured the time of detecting breakable points of an English text (576 characters) and a Japanese text (625 characters).  The values are the average of 1000 runs (on Windows 11 PC, i9-9820X, python 3.9.2).

| text | dict-based(current) | if-based |
|:---:|---:|---:|
| English | 1.98ms | 2.62ms |
| Japanese | 2.32ms | 28.12ms|

If-based fonction was so slow with the Japanese text because plenty of if-sentences must be processed until codes for Japanese characters are found.  The performance for the English text was much better (because codes of English characters are at the top of the Unicode code list), but still slower than the dict-based method.

Next, in line with @alexander-pastukhov, I wrote a module that generates the dict object when imported.  The code was 5562 lines long.  It must be as quick as the current code within the margin of error.

| text | dict-based(current) | dict-based(new) |
|:---:|---:|---:|
| English | 1.98ms | 1.88ms |
| Japanese | 2.32ms | 2.57ms|

The result was that the new code was slightly slower for the Japanese text.  I compared the current and the new dict object and noticed that the current dict object contains fewer entries (333243) than the new one (351115).  This may affected the speed.  I think that the current dict object contains fewer entries because it is based on uniseg-0.7.1 which was released in 2015.  New characters have been added to the Unicode since that time.

In conclusion, I think we can replace the current `linebreaking_class` code with the new code that generates the dict object when importing.  Can I send a pull request?